数式: math_grid のマーカー検出・採番を補助モジュールへ分離する
行末マーカー検出（notag/label）と採番（NumberingMode/assign_numbering 等）を
math_grid/{markers,numbering}.rs へ切り出し、evaluate_grid/evaluate_math_env の
エントリだけを親に残す。振る舞い・外部参照パスは変更なし。

Closes #148
